docsaf/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docsaf"
version = "0.1.0"
description = "DocSAF: Universal Self-Supervised Semantic Attenuation Fields for Document VLM Attacks"
authors = [{name = "DocSAF Team", email = "team@docsaf.ai"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "open-clip-torch>=2.20.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "sentence-transformers>=2.2.0",
    "easyocr>=1.7.0",
    "kornia>=0.7.0",
    "albumentations>=1.3.0",
    "pymupdf>=1.23.0",
    "lpips>=0.1.4",
    "grad-cam>=1.4.8",
    "pytest>=7.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.6.0",
    "scikit-image>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.0.0",
]
paddle = [
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
]

[project.scripts]
docsaf-infer = "src.docsaf.infer_once:main"
docsaf-train = "src.docsaf.train_universal:main"
docsaf-eval = "src.docsaf.eval_harness:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]